backend/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "scentscape-backend"
version = "0.1.0"
description = "ScentScape Backend API - AI-Driven Fragrance Discovery"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "ScentScape Team", email = "dev@scentscape.local"}
]

dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "sqlalchemy>=2.0",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "celery>=5.3.4",
    "redis>=5.0.0",
    "pinecone-client>=2.2.4",
    "neo4j>=5.14.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.1.0",
    "pydantic>=2.0",
    "python-multipart>=0.0.6",
    "slowapi>=0.1.9",
    "sentry-sdk>=1.39.0",
    "httpx>=0.25.0",
    "aiofiles>=23.1.0",
    "prefect>=2.14.0",
    "scrapy>=2.11.0",
    "boto3>=1.26.165",
    "torch-geometric>=2.4.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.8",
    "mypy>=1.7.0",
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.hatch.envs.default]
dependencies = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.8",
    "mypy>=1.7.0",
    "black>=23.12.0",
]
scripts.api = "uvicorn app.main:app --host 0.0.0.0 --port 8000 --reload"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",     # pycodestyle errors
    "W",     # pycodestyle warnings
    "F",     # pyflakes
    "I",     # isort
    "C",     # flake8-comprehensions
    "B",     # flake8-bugbear
    "UP",    # pyupgrade
]
ignore = [
    "E501",  # line too long
]

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
disallow_untyped_calls = false
disallow_incomplete_defs = true
disallow_untyped_defs = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = [
    "neo4j.*",
    "pinecone.*",
    "celery.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--cov=app --cov-report=html --cov-report=term-missing"
